Separate betting in cgame module
