+++ oa_mod/code/cgame/cg_main.c
@@ -297,6 +297,7 @@ vmCvar_t cg_scoreboardSeason;
 vmCvar_t cg_scoreboardAggressive;
 // Utility.
 vmCvar_t cg_scoreboardDefaultSeasonBackup;
+vmCvar_t cg_enableBetting;
 
 typedef struct {
     vmCvar_t* vmCvar;
@@ -514,7 +515,8 @@ static cvarTable_t cvarTable[] = {// bk001129
     {&cg_scoreboardSeason, "cg_scoreboardSeason", "-1", CVAR_ARCHIVE},
     {&cg_scoreboardAggressive, "cg_scoreboardAggressive", "0", CVAR_ARCHIVE},
     // Utility.
-    {&cg_scoreboardDefaultSeasonBackup, "cg_scoreboardDefaultSeasonBackup", "0", CVAR_ARCHIVE}
+    {&cg_scoreboardDefaultSeasonBackup, "cg_scoreboardDefaultSeasonBackup", "0", CVAR_ARCHIVE},
+    {&cg_enableBetting, "g_enableBetting", "1", CVAR_ROM}
 };
 
 static int cvarTableSize = sizeof(cvarTable) / sizeof(cvarTable[0]);
@@ -2218,10 +2220,12 @@ void CG_Init(int serverMessageNum, int serverCommandSequence, int clientNum) {
     s = CG_ConfigString(CS_LEVEL_START_TIME);
     cgs.levelStartTime = atoi(s);
     CG_ParseServerinfo();
-    // oatot
-    trap_SendClientCommand("getActiveBetsSums\n");
-    trap_SendClientCommand("getActiveBets\n");
-    trap_SendClientCommand("getBalance\n");
+    if (cg_enableBetting.integer) {
+        // oatot
+        trap_SendClientCommand("getActiveBetsSums\n");
+        trap_SendClientCommand("getActiveBets\n");
+        trap_SendClientCommand("getBalance\n");
+    }
     cgs.flagsStatus[TEAM_RED].dropped = qfalse;
     cgs.flagsStatus[TEAM_RED].stolen = qfalse;
     // load the new map
